+++ CHANGELOG.md
@@ -2,7 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.3.6] - 2023-10-02
+## [0.4.1] - 2023-10-09
+
+### Added
+
+Added `result_link` field to `Workload` object
+
+### Changed
+
+- `get_workload` now targets v2 of workloads endpoints
+- `result` is built from `result_link` where results are downloaded from temp s3 link
+
+## [0.4.0] - 2023-10-02
 
 ### Added